+++ lib/qc/modules/data_wipe/resources/data_wipe_list_response.g.dart
@@ -35,6 +35,10 @@ DataWipeListItem _$DataWipeListItemFromJson(Map<String, dynamic> json) =>
       (json['sc'] as num?)?.toInt(),
       json['pn'] as String?,
       json['em'] as String?,
+      json['apiname'] as String?,
+      json['imei'] as String?,
+      json['imei2'] as String?,
+      json['sno'] as String?,
     );
 
 Map<String, dynamic> _$DataWipeListItemToJson(DataWipeListItem instance) =>
@@ -46,4 +50,8 @@ Map<String, dynamic> _$DataWipeListItemToJson(DataWipeListItem instance) =>
       'sc': instance.statusCode,
       'pn': instance.productName,
       'em': instance.errorMessage,
+      'apiname': instance.categoryKey,
+      'imei': instance.imei1,
+      'imei2': instance.imei2,
+      'sno': instance.serialNo,
     };